# docs/swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: MockUp API
...
schemes: [sale]
paths:
  /sale/api/v1/AddressMasterData/Zippcode:
    get:
      tags:
        - AddressMasterData
      summary: Get Zipcode List
      produces:
        - text/plain
      operationId: addressMasterData
      parameters: []
      responses:
        401:
          description: Unauthorized
        200:
          description: Success
          schema:
              $ref: '#/definitions/Zipcode'
        500:
          description: Server Error
          schema:
            $ref: "#/definitions/FailedResult"
...
definitions:
...
  FailedResult:
    type: object
    properties:
      code:
        type: string
        # nullable: true
      message:
        type: string
        # nullable: true
  Zipcode:
    type: object
    properties:
      items:
        type: array
        items:
          type: string
          # nullable: true
...